backend/requirements.txt
joblib
numpy
Pillow
scikit-learn